# src/liveChat/server/docs/swagger.yaml
definitions:
  model.ResponseData:
    properties:
      data: {}
      status:
        type: integer
    type: object
  model.UserRequestData:
    properties:
      id:
        type: integer
      password:
        type: string
      repassword:
        type: string
      userName:
        type: string
    type: object
info:
  contact: {}
paths:
  /index:
    get:
      consumes:
      - application/json
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: 获取用户列表
      tags:
      - 用户模块
  /user/deleteUser/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: 用户id
        in: path
        name: id
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResponseData'
      summary: 删除用户
      tags:
      - 用户模块
  /user/login:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户登陆
        in: body
        name: user
        schema:
          $ref: '#/definitions/model.UserRequestData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResponseData'
      summary: 用户登陆
      tags:
      - 用户模块
  /user/updateUser:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户更新信息
        in: body
        name: user
        schema:
          $ref: '#/definitions/model.UserRequestData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResponseData'
      summary: 用户更新
      tags:
      - 用户模块
  /user/userRegistry:
    post:
      consumes:
      - application/json
      parameters:
      - description: 用户注册信息
        in: body
        name: user
        schema:
          $ref: '#/definitions/model.UserRequestData'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.ResponseData'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/model.ResponseData'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/model.ResponseData'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/model.ResponseData'
      summary: 用户注册
      tags:
      - 用户模块
swagger: "2.0"